+++ apps/server/src/device/model.rs
@@ -11,7 +11,7 @@ pub struct Device {
     pub status: u8,
     pub battery: Option<u8>,
     pub location: Option<String>,
-    pub synced_at: Option<u64>,
+    pub synced_at: Option<u32>,
     pub created_at: DateTime<Utc>,
     pub updated_at: Option<DateTime<Utc>>,
 }
